+++ app/models/concerns/coverageable.rb
@@ -41,8 +41,13 @@ module Coverageable
   end
 
   def update_response_count!
-    rcount = non_hidden_responses
-    trcount = non_hidden_total_responses
+    if missing?
+      rcount = 0
+      trcount = 0
+    else
+      rcount = non_hidden_responses
+      trcount = non_hidden_total_responses
+    end
     pcount = compute_percent(rcount, trcount)
     update_columns(
       response_count: rcount,