+++ app/models/subject_event.rb
@@ -47,9 +47,10 @@ class SubjectEvent < ApplicationRecord
           subject_event_id: nil,
           last_user_id: current_user.id,
           last_edited_at: Time.zone.now
-        }, {}, current_user, remote_ip, 'sheet_update', skip_validation: true
+        }, {}, current_user, remote_ip, 'sheet_update', skip_validation: true, skip_callbacks: true
       )
     end
+    update_coverage!
   end
 
   def handoffs?